MB-32108: unlocked_updateStoredValue() to return the StoredValue changed
Subsequent changes for SyncWrites mean that
unlocked_updateStoredValue() can create a new physical StoredValue
object if the SV being modified is Committed and a pending Sync Write
is attempted, as internally the HashTable represents this as a
separate StoredValue.

As such, it is no longer valid to assume that the StoredValue passed
in is the resulting StoredValue. Therefore change the signature of
unlocked_updateStoredValue() to return a struct which includes the
MutationStatus, and a pointer to the StoredValue affected (assuming
update was successful).

Change-Id: I3e97215fbf0400ef4d6b88d3191ef6c94afde2dc
Reviewed-on: http://review.couchbase.org/102388
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>