MB-32108: Introduce HashTable::findFor{Read,Write}
Add two new methods to HashTable for searching for Items -
findForRead() and findForWrite().

As the names imply, these methods search within the HashTable and
return items suitable for reading or writing (modifying). They also
have a simpler signature with sensible default values for WantsDeleted
& TrackReferences given the two use-cases.

In this initial patch the functions only differ in their return types
- findForRead() returns a const StoredValue*, findForWrite a non-const
StoredValue* (alongside the associated HashBucketLock). However in the
next patch when we introduce the concept of accessing the Hashtable
via a Committed or Pending perspective; the behaviour of the functions
will change.

Update existing uses of unlocked_find() / find() to use findForRead()
& findForWrite() as appropriate. A small number of uses of
unlocked_find() remain, where the HashBucketLock has already been
acquired and hence must be passed into the find...() method.

Change-Id: Ie9fb411cfddb1cf7b64c1a7bcf5039ccd41d1695
Reviewed-on: http://review.couchbase.org/102434
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>