Add convenience methods to use char buffers in framebuffer
A lot of in of our unit tests we use a std::string or
char* for the key, but the framebuilder expects a
byte buffer (we've not eliminated one of them yet).

For simplicity start by adding a wrapper method in
the framebuffer to avoid having to cast all over
the place.

Change-Id: I08ce0181b9f592b9791a919f2ccbc2ea9c400456
Reviewed-on: http://review.couchbase.org/102485
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>