MB-26907: Documentation additions for DCP Enable Expiry Output
This patch adds a small .md document to help explain the introduction
of DCP expiry opcode output to other teams. It is located inside our
documentation for ease of linkage use, referring to a couple of packet
information documents; and in turn the control and expiration
documentation link to this.

Change-Id: I6574be9777d72d07950e52af81d4b2ca4b26335d
Reviewed-on: http://review.couchbase.org/102581
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>