Fix missing only-once header defintion and add test for it
Add #pragma once, to make sure that we never re-define
these header files.

Also add python test so that we can made
sure that all headers use a only-once definition
header guard e.g. the presence of #pragma once or

header_define_once_test.py is is pep8 compliant and is compatible
with python2.7 and python3.

Change-Id: Icf4dc0936f3fba5c1c49fc2a1608ab60fdb09d9c
Reviewed-on: http://review.couchbase.org/102527
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>