Refactor: drop iterations from set_huge_impl tests
We're running a dedicated pipeline test later on, so
there isn't much point of doing it in these tests.

Modify the test to send roughly the max document size
(keys and meta is accounted into max size for some backends),
and test the max size when using the quiet ops.

Change-Id: If685dca9781e49d73436dda7a177cf0ba1786188
Reviewed-on: http://review.couchbase.org/102542
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>