MB-32108: Simplify VBQueueItemCtx by adding sensible default values
The majority of uses of VBQueueItemCtx use the same (or very similar)
set of parameters; however currently there is no default constructor
so every use has to specify values for each of the 5 members.

To simplify the usage - and to make it explicit when callers /do/ need
to diverge from the standard usage; add default values for each member
variable, only overriding the ones necessary.

This is in prepararion for adding Durability requirements to
VBQueueItemCtx - we only want to add them where Sync Writes are being
used.

Change-Id: I9d884c74803d0c2d9fc4d1206af905e7b92373ea
Reviewed-on: http://review.couchbase.org/102583
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>