MB-32213: Ensure stale cursors are cleaned-up
It is possible for stale cursors to remain pointing to checkpoints.
For example in ActiveStream::markDiskSnapshot it re-registers a cursor
if we still need to get a memory snapshot.  If when re-registering the
cursor it has moved to a different checkpoint it can result in the
old cursor not being correctly cleaned-up.  These stale cursors can
result in not being able to delete checkpoints.

When a checkpoint used to contain a list of cursor names that currently
point to it, there was a call to a function in
CheckpointManager::removeClosedUnrefCheckpoints that removed these
stale cursors (the function was called
removeInvalidCursorsOnCheckpoint).  However this functionality was
removed in 977643b135afbad273c37f361ad8b8965e233956
(http://review.couchbase.org/#/c/101852/) as we were unware that
stale cursors could arise.

Instead of reverting back to using the removeInvalidCursorsOnCheckpoint
function which required a checkpoint to maintain a list of cursors
pointing to it, this patch instead cleans-up the old cursor prior to
creating a new one with the same name.

Change-Id: I0d5665a0fcee4bd29342d6857af3befa2f5e13ee
Reviewed-on: http://review.couchbase.org/102575
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>