MB-31810: Implement collection in memory high seqno
GSI need the ability to get the highest seqno of a colelction,
just like they already do for a vBucket. To supply the high seqno
for each collection via the VB_GET_ALL_HIGH_SEQNOS DCP command, we
need to add a per collection high seqno counter. This counter should
be incremented when an item is queued for this collection within the
checkpoint manager to the new highest seqno. This counter should be
monotonic and should be initialized from the persistedHighSeqno
counter.

Change-Id: Icb2ae1cfb72be7fd1c3641a399eb876fc46043bc
Reviewed-on: http://review.couchbase.org/101526
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>