MB-32213: Don't increase checkpoint memory for duplicate items
When a duplicate item is added into a checkpoint the old item is
removed and the new item is added.  We don't increase the size of the
checkpoint if the item is a duplicate and the original
item had not yet been persisted.  However, if the original item
had been persisted then we increase the size of the checkpoint.  This
results in an incorrect size for the checkpoint.

This patch addresses the issue by accurately tracking the size of the
checkpoint by increasing the size when an item is added and deleting an
item when it is removed.

Change-Id: Ibc7c5045a13d664d669ba45927539302ec7e2220
Reviewed-on: http://review.couchbase.org/102625
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>