Retry notification of worker threads during shutdown
When moving over to bufferevents calling signalIfIdle
the actual connection may be run at a later time and
not necessarily immediate. This could cause the shutdown
to hang as we've notified the thread, but that notification
what swallowed as part of another "notify_io_complete"
routine causing us to bump back to libevent to wait for
the next event.

Loop and signal the notification pipe until the thread
stops

Change-Id: I170cb39fa91dc47bd3fbe0530a3474a8f5622857
Reviewed-on: http://review.couchbase.org/102719
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>