MB-32336: Negotiate synchronous replication between DcpConsumer & Producer
Add the new dcp_control message 'enable_synchronous_replication' to
indicate that this connection can accept DCP_PREPARE / DCP_COMMIT /
DCP_ABORT messages; and it will respond with SEQNO_ACK messages.

Note that currently the value is just the string 'true'. It is
expected this will be expanded to contain the nodes' UUID in future.

Change-Id: Ia07915041b95c90e5fcb9e5d5706619f99dcb31d
Reviewed-on: http://review.couchbase.org/102738
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>