MB-32099: Change ephemeral eviction policy for cursor dropping tests
This patch attempts to fix the intermittent cv failure, caused by the
pager running and deleting items on an ephemeral bucket. This would
cause an unexpected deletion opcode intermittently. This was
identified on both "test dcp cursor dropping" and "test dcp cursor
dropping (replication)", but not "test dcp cursor dropping backfill".
Seeming as these are similar tests, if we see any CV failures like
this in the future on the backfill test, the same fix should be
applied to it too.

Change-Id: I5a3f6320b6615a0d4866c54dfe30345989689ebb
Reviewed-on: http://review.couchbase.org/102806
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>