MB-31321: Collection connections can't select unsupported buckets
If a connection supports collections, then it should return an
ENotSupported error if the client attempts to select a bucket that
does not support collections. If a client attempts to auth on a
bucket that does not support collections using a connection that does,
then they will be moved to the no bucket, in the same way that they
are if they are not authorized.


Change-Id: Iecc5ce0d6924f8b23c76c73f5db9cdb31cca5148
Reviewed-on: http://review.couchbase.org/102805
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>