Refactor: Add doStreamRequest helper function
DCPTest and subclasses often perform a stream request on the
producer. streamRequest() takes a large number of arguments, many of
which are commonly the same values.

To simplify a number of tests, add a doStreamRequest helper function
which add default parameters for a number of arguments. In many cases
this can be called with only one explicit argument (the producer) for
a basic "give me all the data for vb:0" request.

Change-Id: Ieb59162cac54cee3ae714426f7094a52afc72abe
Reviewed-on: http://review.couchbase.org/102842
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>