MB-26907: Fix deletionCause in Item operators and copy constructor
This patch adds deletionCause to the equivilance operator, the
output operator and the copy constructor inside Item. This involves
adding a to_string method for deletionSource for the output
operator to function. As part of testing this works, an expiration
test for rollback has been added, along with another test to check
that an items values are preserved over a copy.

This fixes MB-32348, an issue seen in xattr output over a DCP
non-xattr stream, as in the process of pruning an item it gets
replicated before it is pruned. The copy contructor originally did
not retain the deletionSource, so the copied item was left
defaulting to 'deleted explicitly', causing the stream to output
a deletion rather than an expiration. By retaining the
deletionSource on replication, this issue is removed.

Change-Id: I8238896ce27092ce490159ca8e7547932da3d7ee
Reviewed-on: http://review.couchbase.org/102839
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>