MB-32336 [SR]: Move SystemEventReplicate::process to ActiveStream
The new DCP messages related to synchronous replication (Prepare,
Abort) are not sent to DCP consumers which do not support sync
replication. This is a per-DCP conneciton property, so the DcpProducer
needs to test each queued item to see if it's relevent to this
stream's config.

There is existing similar functionality for handling SystemEvents in
SystemEventReplicate::process, but that is a static function which
doesn't read any Stream state to make the decision. In preparation for
sending Prepare/Abort only to the appropriate streams, move the
functionality in SystemEventReplicate::process to a method on
ActiveStream. This will subsequently be expanded to handle sync
replication queued ops.

Change-Id: Ibf80a3d5c9b7fa46810663bea23d74dae567717d
Reviewed-on: http://review.couchbase.org/102844
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>