MB-32360: Include collection tombstones in scope filters
When the filter tests if a tombstone matches the filter, it does that
against a set of collection-IDs. That set of collection-IDs though is
built from the VB::Manifest known set of collections, which if a
collection has been fully dropped (and removed from the manifest) will
not feature in the set of collection-IDs.

A collection tombstone must still be included in the DCP scope stream
even if it has fallen out of the VB::Manifest, this commit fixes this.

We have always stored the scope-ID in the tombstone, so we now read it
and make it available to the VB::Filter.

VB::Filter on tombstones doesn't bother using the collection ID set, it
just compares the scope-ID with the filter scope.

Change-Id: I35fd2b2f95c822dca9b788755ba108926156d0b3
Reviewed-on: http://review.couchbase.org/102847
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>