MB-32336 [SR]: Skip pending_sync_write if not supported on DCP stream
DCP connections which do not advertise support for synchronous
replication should not be sent DcpPrepare messages. This was initially
implemented by skipping such event types in DcpProducer::step(),
however that is too late in the sequence as we've already added the
item to the Stream's readyQ.

Instead adopt a similar approach to how Collections events are
filtered - in ActiveStream::shouldProcessItem() check if:

- the Item is a synchronous replication one, and
- if the ActiveStream supports such messages.

If not, then don't process the item.

Also adds initial unit tests for DCP synchronous replication
streams. This required enhancing MockDcpMessageProducers to support
the prepare() method.

Change-Id: I25046a9b979fd96b60b34b925ae0909e46c79110
Reviewed-on: http://review.couchbase.org/102852
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>