MB-32327 [SR]: Bugfix: queue pendingItem if key already committed
In EPVBucket::updateStoredValue(), if the SV to be updated is
Committed and we are updating with a pending Item then it is necessary
to create a new (pending) item in addition to the committed one.

However in this case the newly created (pending) SV wasn't being
queued on the CheckpointManager; the existing pending one was. Fix
this by queuing the correct item; adding a regression test to detect
the issue.

Change-Id: I49b4388e7c80d09200a033a4efff460c53d81af5
Reviewed-on: http://review.couchbase.org/102902
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>