StoredValue: Convert bits2 to plain bitfield
StoredValue::bits2 was initially added as a folly::AtomicBitSet,
however this is unnecessary as the current members of it are always
modified under the HashBucketLock. Additionally a forthcoming patch
needs to change the 'committed' field from one to two bits. Given
AtomicBitSet doesn't allow more than 1 bit to be read / written at
once (so we'd be adding two atomic operations), change the type of
bits2 to a plain bitfield given the extra functionality (and cost) of
AtomicBitSet isn't needed.

Change-Id: Id9bb29da15299f87170898e9ab7599f5483cac34
Reviewed-on: http://review.couchbase.org/102981
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>