MB-32117: Remove verified SyncWrites at commit
With this patch we remove a SyncWrite from the tracked container as soon
as it is committed.

Note that in general a SyncWrite could be referenced by some (only 1 for
now) replica iterator. That is why some enhancement to our tracking
model is necessary, so that it is resilient to the removal of referenced
SyncWrites.

Note that with the new tracking we always keep for a replica:
- the seqno of the last SyncWrite ack'ed
- the last ack'ed seqno
Those are useful for both validating the internal DurabilityMonitor
state and exposing for stats. I will add the latter in a follow-up
patch.

Change-Id: I9495960033420ebd444bf3f4b2046a8299209288
Reviewed-on: http://review.couchbase.org/102986
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>