MB-32381: Fix memory accounting for get/set manifest
Both GET and SET manifest return a std::string which
is allocated inside the engine (increasing mem_used) but
freed from memcached, so mem_used is never decreased, this
can lead to the item pager being forced to run and general
bad times.

The fix is to ensure the engine methods don't pass ownership
of the strings.

get manifest uses ADD_RESPONSE

set manifest just returns an engine_errc and now uses
setErrorContext for the message.

Change-Id: Ic556f4aca4aa5b3ec1d46b883999c7274d6236e9
Reviewed-on: http://review.couchbase.org/102967
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>