StoredDocKey::to_string: don't print size
The size of the key is not really that valuable in most situations,
and leads to somewhat confusing output in things like checkpoint dump
or exception messages - e.g.

    CheckpointManager[0x10de87280] with numItems:2 checkpoints:1
        Checkpoint[0x10deee900] with seqno:{5,5} state:CHECKPOINT_OPEN items:[
        {4,empty,cid:0x1:empty, size:6,102,[m]}
        {5,checkpoint_start,cid:0x1:checkpoint_start, size:11,113,[m]}
        {5,set_vbucket_state,cid:0x1:set_vbucket_state, size:12,114,[m]}
    ]

Note the key is printed here as 'cid:0x1:checkpoint_start, size:11' -
the size:11 part is somewhat ambigious - does that refer to the Item
itself or something else?

Similary when dumping a complete Item, the size is somewhat ambiguous
as we print it later for the value size:

    Item[0x7f45b72b3200] with key:cid:0x0:B1F742DC-11_5, size:e
      value:Blob[0x7f459d131000] with size:31447 age:0 data:
    <cut>

Remove printing size. Also restore the stream back to decimal output
after printing the collection ID.

Change-Id: I27412cdf1760a5f709efd82ca1293c6b443d4796
Reviewed-on: http://review.couchbase.org/103025
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Owen <owend@couchbase.com>