MB-32374: HashTable::insertFromWarmup - don't assume WasClean status
When inserting an item into the HashTable during warmup, if the item
already exists then don't assume the status is always 'WasClean'. This
is because background fetches /are/ permitted during warmup, and as
such the existing item could be a temporary item for bgFetch (or even
a complete item which previolusy completed BGFetch).

This reverts the incorrect assumption added as part of 54425dcb5
(unlocked_updateStoredValue() to return the StoredValue changed).

Change-Id: Ie1b056c8ff4b1cf3ace1e89a51350654c97eb578
Reviewed-on: http://review.couchbase.org/103039
Reviewed-by: Daniel Owen <owend@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>