MB-32394 [SR]: Split 'Committed' state into ViaMutation & ViaPrepare
The StoredValue::committed member is used to distinguish between
Pending & Committed items; both to 'hide' Pending items from view when
not wanted, and also to allow the difference to be passed down to
CheckpointManager & DCP via Item.

However, to support the new DCP_COMMIT message we need to be able to
distinguish between queued_items which were committed as a result of a
Prepare operation, or ones which were 'committed' by a normal mutation
operation - the former should be encoded as DCP_COMMIT, the latter as
DCP_MUTATION.

Given that the queued_items are formed from StoredValue, we hence need
to distinquish StoredValues which were committed via the two
operations previously mentioned.

To achieve this, split Committed into two enum values:

- CommittedViaMutation - A Committed item, which resulted from a
  normal mutation operation.

- CommittedViaPrepare - A Committed item, which results from a
  prepared SyncWrite.

(The exising Pending value stays as-is).

Update the various uses of the enum; and update tests as appropriate.

Change-Id: Id966f53ff18ccdadc1bab20c61068cadb02659f7
Reviewed-on: http://review.couchbase.org/103295
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>