MB-32329: Reset cursor when cursor is removed from map
Although when a cursor is dropped it is removed from the cursor map the
underlying shared_ptr can still be valid due to other uses of the cursor
not yet going out of scope (e.g. ClosedUnrefCheckpointRemoverTask).

Therefore it is possible that the stream's cursor although dropped
does not yet have a reference count of 0 and so attempting to access
the cursor does not result in returning the nullptr.

The problem with this is that if the dcp step function is invoked on the
stream it can result in attempting to do operations on the cursor such
as call getNumItemsForCursor().

To avoid this, when a cursor is dropped we need to set the stream's
cursor pointer to nullptr.

Change-Id: Iab6571cf324df345ffbf90723e9cd4d7d950bb10
Reviewed-on: http://review.couchbase.org/103045
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>