Dump all cookies if when receiving exception in runloop
During OoO we might have multiple cookies available for the
connection (and the exception may not be related to the "first"
cookie that we currently dump in the log.

The log message should also include the full description of the
connection to ease the task of locating the correct connection
(in the case we don't have the hello message in the log message
where the connection description was logged).

Change-Id: I95cbc38ba92a200794a5f7c8e8ad71b03629736c
Reviewed-on: http://review.couchbase.org/103078
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Trond Norbye <trond.norbye@gmail.com>