MB-31266: Count system events
Update the hash-table stats to include items which
are system-event items, for ephemeral buckets this
enables us to provide an more accurate curr_items
statistic which is not inflated by the system
events.

A deleted system event (tombstone) is still tracked
as a system event, no differentiation of alive/deleted.

For persistent buckets more work is needed and is tracked
as MB-26334.

Change-Id: I8ae4d37f1c4b96c4d7c1edc059ea921e17a0ad5a
Reviewed-on: http://review.couchbase.org/103121
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>