Refactor: Allow for running memcached_testapp -v with INFO+
Previously we bumped the verbosity level to DEBUG if -v was
passed to memcached_testapp, but thats the only way to
get memcached_testapp to print to the console.

Unfortunately DEBUG level generates _A LOT_ of output, so this
change puts the output to the console with a single v and if you
want to enable more debug output add a second (or third) v.

Change-Id: I660dc3d24dc4ba280b9072d924b506812c8f73ab
Reviewed-on: http://review.couchbase.org/103252
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>