MB-32394 [SR]: Add queue_op::commit_sync_write
Add a new enumeration value in queue_op, which represents the commit
of a SyncWrite. This will be used by CheckpointManager clients in
subsequent patches to perform the appropriate operation when a
SyncWrite is committed:

- For Flusher, write it to disk as committed.

- For DCP clients which support synchronous replication, transmit as a
  DCP_COMMIT (which will mark the previously prepared mutation as
  committed).

- For DCP clients which do not support synchronous replication,
  transmit as a standard DCP_MUTATION.

Note that toItem now takes an optional durability requirements - if
the CommittedState is pending then this is used to set the correct
durability in the created Item.

Change-Id: I85c7b40a08a8b6fbf82b52d21420eb4748c09b82
Reviewed-on: http://review.couchbase.org/103333
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>