MB-32593: Introduce Memory Tracking Allocator
We intend to move the CheckpointQueue from a stl::list to a std::deque
and want to be able to track any changes in memory overhead.

This patch introduces a memory allocator that allows memory usage to be
tracked.  This is used by the CheckpointQueue to allow accurate
measurement of memory usage.

Change-Id: Ie16a3fcb2534954d0a5590148dc12cb1b8e7e37d
Reviewed-on: http://review.couchbase.org/103342
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>