MB-32535 [SR]: Record client cookie in DurabilityMonitor
The DurabilityMonitor needs to be able to notify the requesting client
connection when a SyncWrite is. completed. To do this we must record
the client's cookie in the DurabilityMonitor, as part of the SyncWrite
object.

Update DurabilityMonitor to hold this additional state; and then
ensure that the client's cookie is passed down into the Durability
Monitor when adding SyncWrites.

Change-Id: I65ad6b23bc903bd04ce49b462dd68b0352e4f229
Reviewed-on: http://review.couchbase.org/103445
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>