MB-32535 [SR]: Notify SyncWrite client when request complete
When a SyncWrite request is complete, notify the client which make the
request so the response can be sent back.

To do this we use notifyIOComplete() method on the engine; however
unlike other existing instances where the notification is done by a
background task; we are calling it from the DurabilityMonitor which
doesn't have direct access to the engine instance. Additionally unlike
existing EWOULDBLOCK notification for stores this isn't a idempotent
operation - if we performed the store() again then that would cause
/another/ mutation to occur. To solve this, we:

1. Define a new SyncWriteCompleted callback object, an instance of
   which is passed into VBucket when it is created. This is typically
   instantiated with a function / lambda from KVBucket which calls
   engine.notifyIOComplete().

2. From VBucket::notifyClientOfCommit() this callback object is
   invoked.

3. Inside EPEngine::storeIfInner() if this is a SyncWrite operation,
   use engineSpecific to record that this is a SyncWrite which has
   successfully been prepared, and that the next call to store()
   should simply return the completed status & CAS of the
   now-committed item.

Change-Id: I38a6dae3aeb98dcedd1ae1ec321db054ff1b7cd7
Reviewed-on: http://review.couchbase.org/103446
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>