MB-32117: Integrate DurabilityMonitor into VBucket
With this patch we essentially switch on the DurabilityMonitor.
If the DurabilityMonitor tracks some pending SyncWrites, at receiving a
proper DCP_SEQNO_ACK the monitor commits the verified SyncWrites by
calling back to VBucket.

This patch introduces also the VBucketDurabilityTest fixture, together
with some first tests that check the status of both HashTable and
CheckpointManager before/after the DurabilityMonitor commit.

Change-Id: Ia0dc1f6473dedce1941150c579a777df2cc3944b
Reviewed-on: http://review.couchbase.org/103326
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>