MB-32573 [SR]: Add Event::Commit, CommitSyncWrite response message
Update the DcpProducer ActiveSteam to convert queued_items of type
commit_sync_write into the appropriate DcpResponse subclass:

- For clients which do not support sync replication they are converted
  into the existing MutationResponse, which is transmitted as a DCP_MUTATION.

- For client which /do/ support sync replication - ie. KV-Engine
  DcpConsumers - convert into a new subclass named CommitSyncWrite,
  which is transmitted as a DCP_COMMIT.

Note this change is diverging from the original design and is using
the mutation's key and not pending_seqno to identify the item to be
committed. The reason for this is that the queued_item
(i.e. ref-counted Item object) would need to store an additional
sequence number (prepared seqno) in addition to the
commit_seqno. Given there's only one type of Item and we don't want to
add 8 bytes to /all/ Item objects - even those which are not
SyncWrites, we instead send the mutations' key in the DCP_COMMIT
message.

This is potentially less effecient in terms of space (seqno is always
8 Bytes, key is variable but likely greater than 8 B).
non-SyncWrites), however is functionally equivilent.

Current status is that the DCP_COMMIT message is transmitted to the
replica node; however it is not processed as the KV-Engine DcpConsumer
has not yet been updated to handle the commit message:

    WARNING 48: Invalid format specified for "DCP_COMMIT" - Status: "Invalid arguments" - Closing connection.

Change-Id: Ib2e28028af4b12ecb1df4e84fd41985aa2e35474
Reviewed-on: http://review.couchbase.org/103456
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Dave Rigby <daver@couchbase.com>