MB-32558: Do not persist pending SyncWrites
Currenlty a pending SyncWrite is persisted as committed, which is
obviously wrong.
This is a temporary change to avoid that pendings are persisted, while
we implement the actual "persist as pending".
With this change we will lose a SyncWrite if the active crashes before
it is committed and persisted, but we prefer that to the current
behaviour.

Change-Id: Id3a3fc9c54dee99560d30c4a6c1b5cb7dfdf94d7
Reviewed-on: http://review.couchbase.org/103453
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>