Refactor: rename queue_dirty_t to current coding standard
Rename enum type to 'QueueDirtyStatus', and the individual enumeration
values to the correct style. Also add a 'Success' prefix to each of
them to make explicit that they correspond to a successful queuing of
an item - a subsequent change will add a new Failing status code.

Also includes a minor restructure of handling SuccessNewItem in
CheckpointManager::queueDirty().

Change-Id: I7dd4aeb94c6920ff18b5210bf5648a2cb4c77cc9
Reviewed-on: http://review.couchbase.org/103498
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>