MB-32574 [SR]: Detect and avoid de-duplicating SyncWrites
When enquing mutations we need to avoid de-duplicating
(i.e. discarding) any existing pending or commited SyncWrites. This is
because we must keep Committed SyncWrites until the subsequent pending
SyncWrite is committed - for a number of reasons but primarily because
a DCP_COMMIT message needs the content of the preceeding DCP_PREPARE
to actually perform the commit (DCP_COMMIT only contains prepared &
commit seqno; not the actual value & metadata being committed).

To handle this, modify Checkpoint::queueDirty() to check for
duplicates which cannot be de-duplicated and return a new status code
FailureDuplicateItem. If the CheckpointManager encounters this case
then it will create a new (empty) checkpoint and retry the
queueDirty().

Change-Id: I2445acc5bda7e8c6985927cd3eb8fcfdcbb80f1f
Reviewed-on: http://review.couchbase.org/103503
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>