MB-32597 [SR]: Propogate StoredValue CommittedState correctly
When an existing item in the HashTable is replaced by a new value the
CommittedState field is not updated. This doesn't affect non-
SyncWrites (as the CommittedState will never diverge from the default
of CommittedViaMutation, but when using SyncWrites this means that
subsequent mutations (e.g. a normal SET) end up with the incorrect
CommittedState of CommitViaPrepare.

Issue is in StoredValue::setValueImpl() which does not copy the
CommittedState from the Item into the StoredValue.

Change-Id: Ic486cf6a9e078430ca68ed56f2e37e94158ad441
Reviewed-on: http://review.couchbase.org/103555
Reviewed-by: Daniel Owen <owend@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>