MB-32598: Reduce the number of times to invoke gdb
The breakpad tests used to invoke gdb several times to inspect
various pieces of the minidump file. Unfortunately the startup
cost of gdb seems to be relatively high so we should try to
reduce the number of times it is invoked.

To work around this we'll invoke gdb a single time, and execute
all of the commands in a single batch. To make it easy to figure
out the result from each command we invoke 'show print pretty'
between each command so that we can easily use the result of
that command to split the output.

Change-Id: Ie2fd605c589753d695caa8fd883cc0a3db13a4f6
Reviewed-on: http://review.couchbase.org/103586
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>